Move g_rename_playlist() to playlist_manager_utils
