+++ foo_ui_columns/playlist_manager_utils.cpp
@@ -103,6 +103,68 @@ bool playlist_format_name_t::titleformat_hook_playlist_t::process_field(
 }
 
 namespace playlist_manager_utils {
+
+class rename_param {
+public:
+    modal_dialog_scope m_scope;
+    pfc::string8* m_text{};
+};
+
+static BOOL CALLBACK RenameProc(HWND wnd, UINT msg, WPARAM wp, LPARAM lp)
+{
+    switch (msg) {
+    case WM_INITDIALOG:
+        SetWindowLongPtr(wnd, DWLP_USER, lp);
+        {
+            auto* ptr = (rename_param*)lp;
+            ptr->m_scope.initialize(FindOwningPopup(wnd));
+            pfc::string_formatter formatter;
+            formatter << R"(Rename playlist: ")" << *ptr->m_text << R"(")";
+            uSetWindowText(wnd, formatter);
+            uSetDlgItemText(wnd, IDC_EDIT, ptr->m_text->get_ptr());
+        }
+        return 1;
+    case WM_COMMAND:
+        switch (wp) {
+        case IDOK: {
+            auto* ptr = (rename_param*)GetWindowLong(wnd, DWLP_USER);
+            uGetDlgItemText(wnd, IDC_EDIT, *ptr->m_text);
+            EndDialog(wnd, 1);
+        } break;
+        case IDCANCEL:
+            EndDialog(wnd, 0);
+            break;
+        }
+        break;
+    case WM_CLOSE:
+        EndDialog(wnd, 0);
+        break;
+    }
+    return 0;
+}
+
+static bool rename_dialog(pfc::string8* text, HWND parent)
+{
+    rename_param param;
+    param.m_text = text;
+    return !!uDialogBox(IDD_RENAME_PLAYLIST, parent, RenameProc, (LPARAM)(&param));
+}
+
+void rename_playlist(size_t index, HWND wnd_parent)
+{
+    static_api_ptr_t<playlist_manager> playlist_api;
+    pfc::string8 temp;
+    if (playlist_api->playlist_get_name(index, temp)) {
+        if (rename_dialog(&temp, wnd_parent)) { // fucko: dialogobx has a messgeloop, someone might have called
+                                                // switcher api funcs in the meanwhile
+            size_t num = playlist_api->get_playlist_count();
+            if (index < num) {
+                playlist_api->playlist_rename(index, temp, temp.length());
+            }
+        }
+    }
+}
+
 bool check_clipboard()
 {
     static_api_ptr_t<ole_interaction> api;
