+++ foo_ui_columns/playlist_manager_utils.h
@@ -50,6 +50,8 @@ public:
 };
 
 namespace playlist_manager_utils {
+void rename_playlist(size_t index, HWND wnd_parent);
+
 bool check_clipboard();
 bool cut(const pfc::list_base_const_t<t_size>& indices);
 bool copy(const pfc::list_base_const_t<t_size>& indices);
